# docs/swagger.yaml
definitions:
  presenter.SMSResp:
    properties:
      id:
        type: integer
      message:
        type: string
      recipient:
        type: string
      status:
        $ref: '#/definitions/presenter.Status'
    type: object
  presenter.SendSMSReq:
    properties:
      message:
        type: string
      recipient:
        type: string
      user_id:
        type: integer
    type: object
  presenter.SendSMSResp:
    properties:
      message:
        type: string
      status:
        $ref: '#/definitions/presenter.Status'
    type: object
  presenter.Status:
    enum:
    - pending
    - success
    - failed
    type: string
    x-enum-varnames:
    - Pending
    - Success
    - Failed
info:
  contact: {}
paths:
  /sms/{id}:
    get:
      consumes:
      - application/json
      description: Get a SMS message by ID
      parameters:
      - description: SMS ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.SMSResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get SMS message
      tags:
      - SMS
  /sms/send:
    post:
      consumes:
      - application/json
      description: Send a SMS message
      parameters:
      - description: Send SMS request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/presenter.SendSMSReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.SendSMSResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send SMS message
      tags:
      - SMS
swagger: "2.0"